Extract shared OAuthIntegrationBase for GoogleDrive/Office365/Nextcloud services
GoogleDriveService, Office365Service, and NextcloudService each hand-rolled
an identical OAuth token-lifecycle (callback URL resolution, provider-config
loading, token store/refresh/expiry, and the 401-retry-once API request
wrapper). Extracted that into OAuthIntegrationBase; each service now keeps
only its provider-specific auth URL/token exchange/refresh logic.

Also fixes two divergence bugs caught along the way:
- Office365Service.listSharePointDrives logged `error: e` (undefined
  variable) instead of the caught error, throwing inside its own catch
  handler whenever a followed-site drive lookup failed.
- GoogleDriveService._buildCallbackUrl read X-Forwarded-* headers directly
  instead of going through getForwardedProto/getForwardedHost (which
  collapse multi-value proxy chains to the trusted value), unlike
  Office365Service and NextcloudService.

Removed Office365Service's dead scope-invalidation block, which logged a
warning about legacy admin-consent scopes but never actually invalidated
anything.

Fixes #1742
